+++ bot/misc/__init__.py
@@ -1,1 +1,1 @@
-from .router import Router
+"""This is the package needed for modules that change as features are added"""
