+++ bot/utils/__init__.py
@@ -1,0 +1,2 @@
+"""This package contains modules to simplify the code"""
+from .router import Router
